agentic_patterns/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-patterns"
version = "0.1.0"
description = "Implementation of core agentic patterns from scratch"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "groq>=0.9.0",
    "python-dotenv>=1.0.0",
    "colorama>=0.4.6",
]

[tool.setuptools.packages.find]
where = ["src"]
